move SetWorkerDaemonsetAsDesired into daemonset.go
