--- config/worker/kustomization.yaml
@@ -1,8 +1,0 @@
-apiVersion: kustomize.config.k8s.io/v1beta1
-kind: Kustomization
-
-resources:
-- service_account.yaml
-- role.yaml
-- rolebinding.yaml
-
