--- config/worker/rolebinding.yaml
@@ -1,13 +1,0 @@
-apiVersion: rbac.authorization.k8s.io/v1
-kind: RoleBinding
-metadata:
-  name: worker
-roleRef:
-  apiGroup: rbac.authorization.k8s.io
-  kind: Role
-  name: worker
-subjects:
-- kind: ServiceAccount
-  name: nfd-worker
-  namespace: node-feature-discovery-operator
-
