+++ internal/daemonset/worker.go
@@ -17,71 +17,45 @@ limitations under the License.
 package daemonset
 
 import (
-	"context"
-
-	appsv1 "k8s.io/api/apps/v1"
 	corev1 "k8s.io/api/core/v1"
-	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
 
-	"sigs.k8s.io/controller-runtime/pkg/controller/controllerutil"
-	nfdv1 "sigs.k8s.io/node-feature-discovery-operator/api/v1"
+	"k8s.io/utils/ptr"
 )
 
 //go:generate mockgen -source=worker.go -package=daemonset -destination=mock_worker.go DaemonsetAPI
 
-func getWorkerEnvs(envVars map[string]string) *[]corev1.EnvVar {
-
-	var obj []corev1.EnvVar
-
-	for k, v := range envVars {
-		obj = append(obj,
-			corev1.EnvVar{
-				Name: k,
-				ValueFrom: &corev1.EnvVarSource{
-					FieldRef: &corev1.ObjectFieldSelector{
-						FieldPath: v,
-					},
-				},
-			})
-	}
-
-	return &obj
-}
-
-func getWorkerAffinity(terms map[string]string) *corev1.Affinity {
-
-	var obj []corev1.NodeSelectorTerm
-
-	for k, v := range terms {
-		obj = append(obj, corev1.NodeSelectorTerm{
-			MatchExpressions: []corev1.NodeSelectorRequirement{
-				{
-					Key:      k,
-					Operator: corev1.NodeSelectorOperator(v),
-				},
-			},
-		})
-	}
-
+func getWorkerAffinity() *corev1.Affinity {
 	return &corev1.Affinity{
 		NodeAffinity: &corev1.NodeAffinity{
 			RequiredDuringSchedulingIgnoredDuringExecution: &corev1.NodeSelector{
-				NodeSelectorTerms: obj,
+				NodeSelectorTerms: []corev1.NodeSelectorTerm{
+					{
+						MatchExpressions: []corev1.NodeSelectorRequirement{
+							{
+								Key:      "node-role.kubernetes.io/master",
+								Operator: "DoesNotExist",
+							},
+						},
+					},
+					{
+						MatchExpressions: []corev1.NodeSelectorRequirement{
+							{
+								Key:      "node-role.kubernetes.io/node",
+								Operator: "Exists",
+							},
+						},
+					},
+				},
 			},
 		},
 	}
-
-}
-
-func BoolPointer(b bool) *bool {
-	return &b
 }
 
 func getWorkerSecurityContext() *corev1.SecurityContext {
 	return &corev1.SecurityContext{
-		ReadOnlyRootFilesystem:   BoolPointer(true),
-		RunAsNonRoot:             BoolPointer(true),
-		AllowPrivilegeEscalation: BoolPointer(false),
+		ReadOnlyRootFilesystem:   ptr.To(true),
+		RunAsNonRoot:             ptr.To(true),
+		AllowPrivilegeEscalation: ptr.To(false),
 		SeccompProfile: &corev1.SeccompProfile{
 			Type: "RuntimeDefault",
 		},
@@ -195,52 +169,4 @@ func getWorkerVolumes() []corev1.Volume {
 
 func getWorkerLabelsAForApp(name string) map[string]string {
 	return map[string]string{"app": name}
-}
-
-func (d *daemonset) SetWorkerDaemonsetAsDesired(ctx context.Context, nfdInstance *nfdv1.NodeFeatureDiscovery, workerDS *appsv1.DaemonSet) error {
-	workerDS.ObjectMeta.Labels = map[string]string{"app": "nfd"}
-
-	envVars := map[string]string{"NODE_NAME": "spec.nodeName"}
-
-	nodeaffinities := map[string]string{
-		"node-role.kubernetes.io/master": "DoesNotExist",
-		"node-role.kubernetes.io/node":   "Exists",
-	}
-
-	workerDS.Spec = appsv1.DaemonSetSpec{
-		Selector: &metav1.LabelSelector{
-			MatchLabels: getWorkerLabelsAForApp("nfd-worker"),
-		},
-		Template: corev1.PodTemplateSpec{
-			ObjectMeta: metav1.ObjectMeta{
-				Labels: getWorkerLabelsAForApp("nfd-worker"),
-			},
-			Spec: corev1.PodSpec{
-				Tolerations: []corev1.Toleration{
-					{
-						Operator: "Exists",
-						Effect:   "NoSchedule",
-					},
-				},
-				Affinity: getWorkerAffinity(nodeaffinities),
-
-				ServiceAccountName: "nfd-worker",
-				DNSPolicy:          corev1.DNSClusterFirstWithHostNet,
-				Containers: []corev1.Container{
-					{
-						Env:             *getWorkerEnvs(envVars),
-						Image:           nfdInstance.Spec.Operand.ImagePath(),
-						Name:            "nfd-worker",
-						Command:         []string{"nfd-worker"},
-						Args:            []string{"--server=nfd-master:$(NFD_MASTER_SERVICE_PORT)"},
-						VolumeMounts:    *getWorkerVolumeMounts(),
-						ImagePullPolicy: getImagePullPolicy(nfdInstance),
-						SecurityContext: getWorkerSecurityContext(),
-					},
-				},
-				Volumes: getWorkerVolumes(),
-			},
-		},
-	}
-	return controllerutil.SetControllerReference(nfdInstance, workerDS, d.scheme)
 }
